Move main loop to dedicated module
